Refactor out gpx internals
